requirements.txt
customtkinter==5.2.0
Pillow==10.0.1
requests==2.31.0
CTkTable==0.8
CTkMessagebox==2.5